move parameters handling functions to deepquest.__init__
this allows them to be executed by __init__ without importing any deepquest modules
to allow set_gpu_id to run before import
